requirements.txt
numpy==1.21.1
pillow==10.2.0
opencv-python==4.9.0.80
pyautogui==0.9.54